[feat/lock]: 잠금화면 처리 기능 개발

+++ domain/challenge/src/main/java/com/hmh/hamyeonham/challenge/model/Apps.kt
@@ -5,6 +5,6 @@ data class Apps(
 ) {
     data class App(
         val appCode: String,
-        val goalTime: Int
+        val goalTime: Long
     )
 }
